pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ollama-telegram-bot"
version = "0.0.9"
description = "Open-source Telegram bot for Ollama"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "python-telegram-bot>=21.10,<22",
  "httpx>=0.27,<0.29",
  "pypdf>=5.3,<6",
  "python-docx>=1.1,<2",
  "openpyxl>=3.1,<4",
  "python-dotenv>=1.0,<2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "ruff>=0.9,<1",
  "mypy>=1.14,<2",
]
